pyproject.toml
[project]
name = "search-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.2",
    "mcp>=1.26.0",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.3",
]

[project.scripts]
search-mcp = "main:run"
